Restructure the Game.tsx file and add the new PlayerGameView.tsx file.
